Allow completer to rewrite the line
Originally the completer could only communicate the completions from the cursor point onward, and the basis for its completion, so only suffixes of the current input can be completed. With this change, that is preserved, but now the completer can also rewrite previous tokens to allow a more assistive input mode.

Additionally, we add a new value passed from the completion callback, so instead of `[[ completions ], completeOn ]`, we can use `{completions, completeOn, replace }` and extend it with further values in the future.